% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcame_aipe.R
\name{print.dcame_aipe}
\alias{print.dcame_aipe}
\title{Print method for \code{dcame_aipe} results}
\usage{
\method{print}{dcame_aipe}(x, ...)
}
\arguments{
\item{x}{a \code{dcame_aipe} object returned by \code{\link[=dcame_aipe]{dcame_aipe()}}.}

\item{...}{unused}
}
\description{
Print method for \code{dcame_aipe} results
}
