% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcame_aipe.R
\name{print.dcame_aipe_grob}
\alias{print.dcame_aipe_grob}
\title{Print the combined plot grob, redrawing the figure}
\usage{
\method{print}{dcame_aipe_grob}(x, ...)
}
\arguments{
\item{x}{a \code{dcame_aipe_grob} object}

\item{...}{unused}
}
\description{
Print the combined plot grob, redrawing the figure
}
